% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_sim.R
\name{cpue.sim}
\alias{cpue.sim}
\title{To simulate spatially coorelated cpue data over the sampling area}
\usage{
cpue.sim(N, sim_locs, variog_fit, q_bar, sd_q, n_sim, A, n_max = 100)
}
\arguments{
\item{N}{The "true" number of fish}

\item{sim_locs}{A data.frame representation of the coordinate locations over which cpue data will be simulated}

\item{variog_fit}{A vector containing estimates of the nugget, partial sill, and range parameter for the exponential variogram. The parameter estimates must be given in that order.}

\item{q_bar}{An estimate of the catchability coefficient}

\item{sd_q}{The estimated standard deviation of the catchability coefficient}

\item{n_sim}{The number of simulations}

\item{A}{The area of the lake (in XX)}

\item{n_max}{For local kriging -- he number of nearest observations that should be used for kriging simulation. see help(gstat)}
}
\value{
Returns a data.frame containing the coordinate locations simulated cpue values
}
\description{
To simulate spatially coorelated cpue data over the sampling area
}
