% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_sample_area.R
\name{crop.sample.area}
\alias{crop.sample.area}
\title{To determine which points are in the sampling area}
\usage{
crop.sample.area(extent, sp, dist, units)
}
\arguments{
\item{sp}{A SpatialPolygons representation of the section of the lake that can be sampled}

\item{dist}{Numerical argument to control the within transect set spacing}
}
\value{
Returns a data.frame with the coordinates over which data should be simulated
}
\description{
To determine which points are in the sampling area
}
