% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_variog.R
\name{fit.variog}
\alias{fit.variog}
\title{To estimate the degree of spatial autocoorelation in cpue count data using an exponential variogram}
\usage{
fit.variog(count_data)
}
\arguments{
\item{count_data}{A SpatialPointsDataFrame storing the cpue data}
}
\value{
Returns a vector containing estimates of the nugget, partial sill, and range parameter respectively
}
\description{
To estimate the degree of spatial autocoorelation in cpue count data using an exponential variogram
}
