% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage_sample.R
\name{two.stage.sample}
\alias{two.stage.sample}
\title{Function to sample from the simulated cpue according to a predefined sampling design}
\usage{
two.stage.sample(cpue_sim, n_sets)
}
\arguments{
\item{cpue_sim}{A data.frame containing the simulated cpue values and coordinate locations}

\item{n_sets}{The total number of sets}
}
\value{
Returns a list of data.frames where each contains the transect #, set within transect #, and cpue value
}
\description{
Function to sample from the simulated cpue according to a predefined sampling design
}
